# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Sample Booking API
  description: API Documentation in Markdown.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    variables: {}
paths:
...
  /administrators/hotels/:hotelId:
    put:
      summary: Update hotel
      tags:
        - Administrators
      description: Update selected hotel
      operationId: UpdateHotel
      parameters: []
      requestBody:
        description: ''
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  required: true
                rating:
                  type: integer
                  required: true
                main_pic:
                  type: image
                  required: true
                second_pic:
                  type: image
                  required: true
                main_picture:
                  type: string
                  description: File id
                  required: true
                second_picture:
                  type: string
                  description: File id
                  required: true
                short_address:
                  type: string
                  required: true
                address:
                  type: string
                  required: true
                location:
                  type: string
                  required: true
                type:
                  type: string
                  required: true
        required: true
      responses:
        '200':
          description: 'Success'
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/hotelCreateSuccess'
        '400':
          description: Hotel id not found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User is not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete hotel
      tags:
        - Administrators
      description: Delete a hotel
      operationId: DeleteHotel
      parameters: []
      responses:
        '200':
          description: 'Success'
          headers: {}
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  result:
                    type: object
                    properties:
                      message:
                        type: string
                        default: 'Hotel deleted'
        '400':
          description: Hotel id not found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Admin is not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    hotelCreateSuccess:
      title: Create hotel success
      type: object
      properties:
        ok:
          type: boolean
          default: true
        result:
          type: object
          properties:
            message:
              type: string
              default: 'Hotel created'
            data:
              type: array
              items:
                $ref: '#/components/schemas/Hotel'
              description: ''
    Hotel:
      title: Hotel
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        type:
          type: string
        rating:
          type: number
        main_picture:
          type: string
        second_picture:
          type: string
        short_address:
          type: string
        address:
          type: string
        location:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    Error:
      title: Error
      type: object
      properties:
        ok:
          type: boolean
          default: false
        result:
          type: object
          properties:
            message:
              type: string
            data:
              type: object